refactor(v0.76.0): commands.py 2441L → 13-file 패키지로 분할 (Tier 3 #9) (#907)
Tier 3 God Object 분할 시리즈 마지막. 사용자-facing CLI 슬래시
명령 라우터 (모든 /key, /model, /auth, /login, /cost, /skills, /mcp,
/compact, /clear, /resume, /apply, /context, /tasks, /trigger 진입점)인
2441 LOC monolith를 명령 family별 13개 모듈로 분리.

새 패키지 구조 — core/cli/commands/:
  __init__.py    148L  53개 심볼 re-export (외부 29곳 import 변경 0)
  _state.py      201L  ModelProfile, MODEL_PROFILES, COMMAND_MAP,
                       _conversation_ctx, install_domain_commands,
                       show_help, resolve_action, _get_profile_store
  key.py         211L  cmd_key, _seed_payg_plan_from_key, _persist_auth_state,
                       _check_provider_key
  model.py       204L  _apply_model, _interactive_model_picker, cmd_model
  auth.py        316L  _auth_login_status, cmd_auth, _auth_add_interactive,
                       _sync_oauth_profile_after_login
  mcp.py         114L  cmd_mcp, _mcp_add
  skills.py      200L  cmd_skills, cmd_skill_invoke, _skills_add
  cost.py        230L  cmd_cost, _budget_bar, _get/_set_cost_budget
  session.py     418L  cmd_resume, cmd_apply, cmd_context, cmd_compact, cmd_clear
  tasks.py        84L  cmd_tasks
  trigger.py      50L  cmd_trigger
  login.py       655L  cmd_login + 9 _login_* helpers (최대 leaf, 의도적 보존)

최대 단일 파일: 2441 → 655 LOC (-73%, Tier 3 시리즈 최대 절대 감소).

테스트 monkeypatch 호환성: 28개 @patch("core.cli.commands.X") 사이트
보존을 위해 sub-module 함수가 deferred `from core.cli import commands as _pkg`
조회 후 `_pkg.console.print(...)`, `_pkg._upsert_env(...)` 등 사용 —
agentic_ui/tool_executor 분할의 검증된 패턴 재적용.

플러그인 호환성: plugins/game_ip/__init__.py 의 COMMAND_MAP.update()
mutation 그대로 작동 (re-export는 reference, 동일 dict 객체).

공개 API 변경 0건. 외부 import 29곳 (core/ 18, plugins/ 2, tests/ 9)
수정 불필요 — 53개 심볼 모두 __init__.py re-export 통과.

Quality Gates:
- ruff/ruff-format/mypy 0 errors (321 source files, +11)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- 명령-related 8 테스트 파일 (commands, e2e, compact_clear, session_resume,
  login_command, oauth_login, manage_login_tool, provider_label_consistency)
  155 passed
- E2E Cowboy Bebop A (68.4) 변동 없음
- pyproject.toml import-linter 변경 0 (rule이 leaf path 그대로 resolve)

남은 Tier 3 후보 1개 — cli/__init__.py(1889L, 0 class).
